refactor(monolith): extract normalize_string to utils.text and delegate from monolith
